210618 columnDefinition 수정 및 금액 자료형 변경 - Entity column에 columnDefinition 수정 - Entity 금액 컬럼 Int > Long 으로 수정 - 주석 제거

+++ src/main/java/com/app/dto/BudgetListDto.java
@@ -19,7 +19,7 @@ public class BudgetListDto {
 
     public Budget saveExpenditureBudget(UserDto userDto){
         return Budget.builder()
-                .budgetAmount(Integer.parseInt(budgetAmount))
+                .budgetAmount(Long.parseLong(budgetAmount))
                 .budgetDate(budgetDate)
                 .largeCategoryId(largeCategoryId)
                 .user(userDto.toEntity())
